Add submitCarToStore function
On submiting the form an object car with all necessary params
is created and a function creteCar
is called. To this function the car object is passed and using axios
the object is sent to the api. Method post is used here.
So the new created object can be saved to the database.
In case the post method is successfull action addCar
with passed param the car object is dispatched.