Create DeleteItem page
In order to delete an item a
selector is used to act a drop down menu. After a car is selected
onSubmit event a deleteCar
method is called. In this methos using axios and delete method
a action with parameter(the id of the selected car) is dispatched.

By clicking the close icon the user is redirect to the main page.