Create  and style Reserve page
All the cities from the api are loadied using useEffect, so a user
can select a city in the form from
the drop-down menu.
The data filled in in the form
is send to set the state. And when is submitted an action is dispated
with the neccessary parameters.

Add styling to the page